requirements-dev.txt
torch==2.2.1
torchvision==0.17.1
hydra-core==1.3.2
mlflow==2.11.1
scikit-learn==1.4.1.post1
numpy==1.26.4
matplotlib==3.8.3
matplotlib-inline==0.1.6
kserve==0.12.0
pillow==10.2.0
